+++ lib/types.ts
@@ -1,33 +1,40 @@
 import type { z } from 'zod/v4'
 import type { $strict } from 'zod/v4/core'
 
-type Prev = [never, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]
+/* --------- CORE --------- */
 
-export type SupportedZodPrimitives =
-	| z.ZodString
-	| z.ZodNumber
-	| z.ZodBoolean
-	| z.ZodUndefined
-	| z.ZodNull
-	| z.ZodLiteral<string>
-	| z.ZodEnum<Record<string, string>>
+type DepthMap = [never, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
 
-export type SupportedZodTypes<Depth extends number = 20> = Depth extends 0
+type MAX_DEPTH = 10
+
+export type $SupportedZodTypes<Depth extends number = MAX_DEPTH> = Depth extends 0
 	? never
 	:
-			| SupportedZodPrimitives
-			| z.ZodObject<Record<string, SupportedZodTypes<Prev[Depth]>>>
-			| z.ZodArray<SupportedZodTypes<Prev[Depth]>>
-			| z.ZodUnion<readonly SupportedZodTypes<Prev[Depth]>[]>
+			| z.ZodString
+			| z.ZodNumber
+			| z.ZodBoolean
+			| z.ZodUndefined
+			| z.ZodNull
+			| z.ZodLiteral<string>
+			| z.ZodEnum<Record<string, string>>
+			| z.ZodObject<Record<string, $SupportedZodTypes>>
+			| z.ZodArray<$SupportedZodTypes>
+			// LIMIT RECURSIVE DEPTH ON UNIONS (TS2589)
+			| z.ZodUnion<readonly $SupportedZodTypes<DepthMap[Depth]>[]>
 
-export type SupportedZodCompoundTypes = Exclude<SupportedZodTypes, SupportedZodPrimitives>
+export type SupportedZodTypes = $SupportedZodTypes
 
 export type Node = {
 	input: Record<string, SupportedZodTypes>
 	output: SupportedZodTypes
 }
 
-export type ShapeOf<T extends SupportedZodTypes> = T extends z.ZodString
+export type CustomCompatibility = {
+	type: SupportedZodTypes
+	compatibilities: SupportedZodTypes[]
+}
+
+type ShapeOf<T extends SupportedZodTypes> = T extends z.ZodString
 	? 'string'
 	: T extends z.ZodNumber
 		? 'number'
@@ -45,46 +52,93 @@ export type ShapeOf<T extends SupportedZodTypes> = T extends z.ZodString
 								? { [K in keyof InnerShape]: ShapeOf<InnerShape[K]> }
 								: T extends z.ZodArray<infer InnerShape extends SupportedZodTypes>
 									? ShapeOf<InnerShape>[]
-									: T extends z.ZodUnion<infer Union extends SupportedZodTypes[]>
+									: T extends z.ZodUnion<infer Union extends readonly SupportedZodTypes[]>
 										? ShapeOf<Union[number]>
 										: never
 
-type Compatable<
-	Type1 extends SupportedZodTypes,
-	Type2 extends SupportedZodTypes
-> = ShapeOf<Type1> extends ShapeOf<Type2> ? true : false
+type IsCompatible<
+	Out extends SupportedZodTypes,
+	In extends SupportedZodTypes
+> = ShapeOf<Out> extends ShapeOf<In> ? true : false
 
-export type NodeCompatability<
+/* --------- COMPATABILITIES --------- */
+
+type $NodeCompatibilities<
 	Type extends SupportedZodTypes,
-	Nodes extends Record<string, Node>
+	Nodes extends Record<string, Node>,
+	Strict extends boolean,
+	CustomCompatabilities extends CustomCompatibility[]
 > = {
-	[K in keyof Nodes]: Compatable<Nodes[K]['output'], Type> extends true
-		? NodeDefinition<K & string, Nodes[K]['input'], Nodes>
+	[K in keyof Nodes]: IsCompatible<Nodes[K]['output'], Type> extends true
+		? NodeDefinition<K & string, Nodes[K]['input'], Nodes, Strict, CustomCompatabilities>
 		: never
 }[keyof Nodes][]
 
-type NodeDefinition<
+type $InnerCompatibilities<
+	Type extends SupportedZodTypes,
+	Nodes extends Record<string, Node>,
+	Strict extends boolean,
+	CustomCompatabilities extends CustomCompatibility[]
+> = Type extends z.ZodArray<infer Element extends SupportedZodTypes>
+	? [z.ZodArray<Compatibilities<Element, Nodes, Strict, CustomCompatabilities>>]
+	: Type extends z.ZodObject<infer Fields extends Record<string, SupportedZodTypes>>
+		? [
+				z.ZodObject<{
+					[K in keyof Fields]: Compatibilities<Fields[K], Nodes, Strict, CustomCompatabilities>
+				}>
+			]
+		: Type extends z.ZodUnion<infer Options extends readonly SupportedZodTypes[]>
+			? [
+					z.ZodUnion<{
+						[I in keyof Options]: Compatibilities<Options[I], Nodes, Strict, CustomCompatabilities>
+					}>
+				]
+			: []
+
+type $AdditionalCompatabilities<
+	Type extends SupportedZodTypes,
+	AdditionalCompatabilities extends CustomCompatibility[]
+> = [
+	AdditionalCompatabilities[number] extends infer AdditionalCompatability
+		? AdditionalCompatability extends CustomCompatibility
+			? IsCompatible<AdditionalCompatability['type'], Type> extends true
+				? AdditionalCompatability['compatibilities'][number]
+				: never
+			: never
+		: never
+]
+
+export type Compatibilities<
+	Type extends SupportedZodTypes,
+	Nodes extends Record<string, Node>,
+	Strict extends boolean,
+	AdditionalCompatabilities extends CustomCompatibility[]
+> = z.ZodUnion<
+	[
+		...$NodeCompatibilities<Type, Nodes, Strict, AdditionalCompatabilities>,
+		...$InnerCompatibilities<Type, Nodes, Strict, AdditionalCompatabilities>,
+		...$AdditionalCompatabilities<Type, AdditionalCompatabilities>,
+		...(Strict extends true ? [] : [Type])
+	]
+>
+
+/* --------- MAIN CAST --------- */
+
+export type NodeDefinition<
 	Name extends string,
 	Input extends Record<string, SupportedZodTypes>,
-	Nodes extends Record<string, Node>
+	Nodes extends Record<string, Node>,
+	Strict extends boolean,
+	AdditionalCompatabilities extends CustomCompatibility[]
 > = z.ZodObject<
 	{
 		node: z.ZodLiteral<Name>
 		input: z.ZodObject<
 			{
-				[K in keyof Input]: z.ZodUnion<NodeCompatability<Input[K], Nodes>>
+				[K in keyof Input]: Compatibilities<Input[K], Nodes, Strict, AdditionalCompatabilities>
 			},
 			$strict
 		>
 	},
 	$strict
 >
-
-export type Definition<
-	Name extends string = string,
-	Input extends Record<string, SupportedZodTypes> = Record<string, SupportedZodTypes>,
-	Nodes extends Record<string, Node> = Record<string, Node>
-> = {
-	definition: NodeDefinition<Name, Input, Nodes>
-	output: SupportedZodTypes
-}
